Refactor pipeline and update project documentation
Refactored the pipeline execution mechanism by introducing a `PipelineBehaviorFactory` for dynamic behavior resolution and optimizing `CommandInvokerFactory` with caching. Updated `PipelineExecutor` to integrate the new factory and ensure proper chaining of behaviors. Removed redundant interfaces and files for clarity. #39

Significantly enhanced the `README.md` to provide a detailed overview of the project's goals, architecture, and current status. Improved maintainability by cleaning up unused code and comments.
