+++ src/Rewardly.Application/Abstractions/ICommandInvokerFactory.cs
@@ -1,0 +1,14 @@
+﻿namespace Rewardly.Application.Abstractions;
+
+/// <summary>
+/// 
+/// </summary>
+public interface ICommandInvokerFactory
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="command"></param>
+    /// <returns></returns>
+    ICommandInvoker Create(ICommandBase command);
+}
