+++ src/Rewardly.Application/Bus/CommandBus.cs
@@ -2,7 +2,7 @@
 
 namespace Rewardly.Application.Bus;
 
-public class CommandBus : ICommandBus
+public sealed class CommandBus : ICommandBus
 {
     private readonly IPipelineExecutor _pipeline;
 
@@ -11,7 +11,7 @@ public class CommandBus : ICommandBus
         _pipeline = pipeline;
     }
 
-    public async Task<TResponse> SendAsync<TResponse>(ICommand<TResponse> command, CancellationToken cancellationToken)
+    public async Task<TResponse> SendAsync<TResponse>(ICommand<TResponse> command, CancellationToken cancellationToken) 
     {
         return await _pipeline.ExecuteAsync(command, cancellationToken);
     }
