+++ src/Rewardly.Application/Interfaces/Pipeline/IPipelineBehavior.cs
@@ -1,6 +1,13 @@
 ﻿namespace Rewardly.Application.Interfaces.Pipeline;
 
 /// <summary>
+/// Representa o próximo manipulador na cadeia de execução do pipeline.
+/// </summary>
+/// <typeparam name="TResponse">Tipo da resposta produzida pelo manipulador.</typeparam>
+/// <returns>Tarefa assíncrona que retorna a resposta do próximo passo do pipeline.</returns>
+public delegate Task<TResponse> RequestHandlerDelegate<TResponse>();
+
+/// <summary>
 /// Define um comportamento de pipeline executado em torno do manipulador de comandos.
 /// </summary>
 /// <typeparam name="TRequest">Tipo do comando processado no pipeline.</typeparam>
