+++ src/Rewardly.Application/IoC/ApplicationModule.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using Rewardly.Application.Abstractions;
 using Rewardly.Application.Bus;
 using Rewardly.Application.Interfaces.Pipeline;
 using Rewardly.Application.Services.v1;
