+++ src/Rewardly.Application/Pipeline/PipelineExecutor.cs
@@ -1,36 +1,36 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Rewardly.Application.Abstractions;
 using Rewardly.Application.Interfaces.Pipeline;
 
 namespace Rewardly.Application.Pipeline;
 
 public sealed class PipelineExecutor : IPipelineExecutor
 {
-    private readonly IServiceProvider _serviceProvider;
     private readonly ICommandInvokerFactory _commandInvokerFactory;
+    private readonly IPipelineBehaviorFactory _behaviorFactory;
 
-    public PipelineExecutor(IServiceProvider serviceProvider, ICommandInvokerFactory commandInvokerFactory)
+    public PipelineExecutor(ICommandInvokerFactory commandInvokerFactory, IPipelineBehaviorFactory behaviorFactory)
     {
-        _serviceProvider = serviceProvider;
         _commandInvokerFactory = commandInvokerFactory;
+        _behaviorFactory = behaviorFactory;
     }
 
     public async Task<TResponse> ExecuteAsync<TResponse>(ICommand<TResponse> command, CancellationToken cancellationToken)
     {
         ICommandInvoker? invoker = _commandInvokerFactory.Create(command);
 
+        List<object> behaviors = _behaviorFactory.Create(command).Reverse().ToList();
+
         RequestHandlerDelegate<TResponse> next = async () =>
         {
             object? resut = await invoker.InvokeAsync(command, cancellationToken);
-            return (TResponse)resut;
+            return (TResponse)resut!;
         };
 
-        var behaviors = _serviceProvider.GetServices<IPipelineBehavior<ICommand<TResponse>, TResponse>>().Reverse().ToList();
-
-        foreach (var behavior in behaviors)
+        foreach (dynamic behavior in behaviors)
         {
-            var current = next;
+            RequestHandlerDelegate<TResponse> current = next;
 
-            next = () => behavior.HandleAsync(command, current, cancellationToken);
+            next = () => behavior.HandleAsync((dynamic)command, current, cancellationToken);
         }
 
         return await next();
